Migrate burst utils to dedicated script
